% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.R
\name{sum}
\alias{sum}
\title{Hello, world!}
\usage{
sum(x, y)
}
\arguments{
\item{x}{arg1}

\item{y}{arg2
Some useful keyboard shortcuts for package authoring:

  Build and Reload Package:  'Ctrl + Shift + B'
  Check Package:             'Ctrl + Shift + E'
  Test Package:              'Ctrl + Shift + T'}
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\details{
You can learn more about package authoring with RStudio at:

http://r-pkgs.had.co.nz/
}
